WebServiceZonesApi: Updated AddRecord(), DeleteRecord(), and UpdateRecord() to add missing check for permissions when handling the PTR option for A/AAAA records.

+++ DnsServerCore/WebServiceZonesApi.cs
@@ -3872,6 +3872,9 @@ namespace DnsServerCore
                                     if (!createPtrZone)
                                         throw new DnsWebServiceException("No reverse zone available to add PTR record.");
 
+                                    if (!_dnsWebService._authManager.IsPermitted(PermissionSection.Zones, sessionUser, PermissionFlag.Modify))
+                                        throw new DnsWebServiceException("Cannot create reverse zone to add PTR record: access was denied.");
+
                                     string ptrZone = Zone.GetReverseZone(ipAddress, type == DnsResourceRecordType.A ? 24 : 64);
 
                                     reverseZoneInfo = _dnsWebService._dnsServer.AuthZoneManager.CreatePrimaryZone(ptrZone);
@@ -3883,6 +3886,11 @@ namespace DnsServerCore
                                     _dnsWebService._authManager.SetPermission(PermissionSection.Zones, reverseZoneInfo.Name, _dnsWebService._authManager.GetGroup(Group.ADMINISTRATORS), PermissionFlag.ViewModifyDelete);
                                     _dnsWebService._authManager.SetPermission(PermissionSection.Zones, reverseZoneInfo.Name, _dnsWebService._authManager.GetGroup(Group.DNS_ADMINISTRATORS), PermissionFlag.ViewModifyDelete);
                                     _dnsWebService._authManager.SaveConfigFile();
+                                }
+                                else
+                                {
+                                    if (!_dnsWebService._authManager.IsPermitted(PermissionSection.Zones, reverseZoneInfo.Name, sessionUser, PermissionFlag.Modify))
+                                        throw new DnsWebServiceException("Cannot update reverse zone to add PTR record: access was denied.");
                                 }
 
                                 if ((reverseZoneInfo.Type != AuthZoneType.Primary) && (reverseZoneInfo.Type != AuthZoneType.Forwarder))
@@ -4352,9 +4360,12 @@ namespace DnsServerCore
                                     {
                                         if ((ptrRecord.RDATA as DnsPTRRecordData).Domain.Equals(domain, StringComparison.OrdinalIgnoreCase))
                                         {
-                                            //delete PTR record and save reverse zone
-                                            _dnsWebService._dnsServer.AuthZoneManager.DeleteRecord(reverseZoneInfo.Name, ptrDomain, DnsResourceRecordType.PTR, ptrRecord.RDATA);
-                                            _dnsWebService._dnsServer.AuthZoneManager.SaveZoneFile(reverseZoneInfo.Name);
+                                            if (_dnsWebService._authManager.IsPermitted(PermissionSection.Zones, reverseZoneInfo.Name, sessionUser, PermissionFlag.Delete))
+                                            {
+                                                //delete PTR record and save reverse zone
+                                                _dnsWebService._dnsServer.AuthZoneManager.DeleteRecord(reverseZoneInfo.Name, ptrDomain, DnsResourceRecordType.PTR, ptrRecord.RDATA);
+                                                _dnsWebService._dnsServer.AuthZoneManager.SaveZoneFile(reverseZoneInfo.Name);
+                                            }
                                             break;
                                         }
                                     }
@@ -4680,6 +4691,9 @@ namespace DnsServerCore
                                     if (!createPtrZone)
                                         throw new DnsWebServiceException("No reverse zone available to add PTR record.");
 
+                                    if (!_dnsWebService._authManager.IsPermitted(PermissionSection.Zones, sessionUser, PermissionFlag.Modify))
+                                        throw new DnsWebServiceException("Cannot create reverse zone to add PTR record: access was denied.");
+
                                     string ptrZone = Zone.GetReverseZone(newIpAddress, type == DnsResourceRecordType.A ? 24 : 64);
 
                                     newReverseZoneInfo = _dnsWebService._dnsServer.AuthZoneManager.CreatePrimaryZone(ptrZone);
@@ -4692,6 +4706,11 @@ namespace DnsServerCore
                                     _dnsWebService._authManager.SetPermission(PermissionSection.Zones, newReverseZoneInfo.Name, _dnsWebService._authManager.GetGroup(Group.DNS_ADMINISTRATORS), PermissionFlag.ViewModifyDelete);
                                     _dnsWebService._authManager.SaveConfigFile();
                                 }
+                                else
+                                {
+                                    if (!_dnsWebService._authManager.IsPermitted(PermissionSection.Zones, newReverseZoneInfo.Name, sessionUser, PermissionFlag.Modify))
+                                        throw new DnsWebServiceException("Cannot update reverse zone to add PTR record: access was denied.");
+                                }
 
                                 if ((newReverseZoneInfo.Type != AuthZoneType.Primary) && (newReverseZoneInfo.Type != AuthZoneType.Forwarder))
                                     throw new DnsWebServiceException("Reverse zone '" + newReverseZoneInfo.DisplayName + "' is not a primary or forwarder zone.");
@@ -4701,6 +4720,9 @@ namespace DnsServerCore
                                 AuthZoneInfo oldReverseZoneInfo = _dnsWebService._dnsServer.AuthZoneManager.FindAuthZoneInfo(oldPtrDomain);
                                 if ((oldReverseZoneInfo is not null) && ((oldReverseZoneInfo.Type == AuthZoneType.Primary) || (oldReverseZoneInfo.Type == AuthZoneType.Forwarder)))
                                 {
+                                    if (!_dnsWebService._authManager.IsPermitted(PermissionSection.Zones, oldReverseZoneInfo.Name, sessionUser, PermissionFlag.Delete))
+                                        throw new DnsWebServiceException("Cannot update reverse zone to delete existing PTR record: access was denied.");
+
                                     //delete old PTR record if any and save old reverse zone
                                     _dnsWebService._dnsServer.AuthZoneManager.DeleteRecords(oldReverseZoneInfo.Name, oldPtrDomain, DnsResourceRecordType.PTR);
                                     _dnsWebService._dnsServer.AuthZoneManager.SaveZoneFile(oldReverseZoneInfo.Name);
